feat(OHIFv3):Add CobbAngle tool support
