+++ extensions/cornerstone/src/utils/measurementServiceMappings/constants/supportedTools.js
@@ -1,1 +1,1 @@
-export default ['Length', 'EllipticalRoi', 'Bidirectional', 'ArrowAnnotate'];
+export default ['Length', 'EllipticalRoi', 'Bidirectional', 'ArrowAnnotate', 'CobbAngle'];
