+++ platform/i18n/src/locales/en-US/Buttons.json
@@ -6,6 +6,7 @@
   "Brush": "Brush",
   "CINE": "CINE",
   "Cancel": "Cancel",
+  "CobbAngle": "Cobb Angle",
   "Circle": "Circle",
   "Clear": "Clear",
   "Coronal": "Coronal",
